Uninstall improvements

+++ commands/command_uninstall.go
@@ -10,11 +10,13 @@ func uninstallCommand(cmd *cobra.Command, args []string) {
 		Error(err.Error())
 	}
 
-	if localInstall || cfg.InRepo() {
+	if !skipRepoInstall && (localInstall || cfg.InRepo()) {
 		uninstallHooksCommand(cmd, args)
 	}
 
-	if !localInstall {
+	if systemInstall {
+		Print("System Git LFS configuration has been removed.")
+	} else if !localInstall {
 		Print("Global Git LFS configuration has been removed.")
 	}
 }
@@ -32,6 +34,7 @@ func init() {
 	RegisterCommand("uninstall", uninstallCommand, func(cmd *cobra.Command) {
 		cmd.Flags().BoolVarP(&localInstall, "local", "l", false, "Set the Git LFS config for the local Git repository only.")
 		cmd.Flags().BoolVarP(&systemInstall, "system", "", false, "Set the Git LFS config in system-wide scope.")
+		cmd.Flags().BoolVarP(&skipRepoInstall, "skip-repo", "", false, "Skip repo setup, just uninstall global filters.")
 		cmd.AddCommand(NewCommand("hooks", uninstallHooksCommand))
 	})
 }
